backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
motor>=3.3.0  # Async MongoDB driver
pymongo>=4.6.0

# HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Firecrawl
firecrawl-py>=1.0.0

# Environment and config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.0
